Moved the flatten cqc ratings job, test, test data/shcema and updated glue script
