+++ projects/_02_sfc_internal/cqc_ratings/jobs/flatten_cqc_ratings.py
