+++ projects/_02_sfc_internal/unittest_data/sfc_test_file_data.py
@@ -1,8 +1,14 @@
 from dataclasses import dataclass
 from datetime import date
 
+from utils.column_names.raw_data_files.cqc_location_api_columns import (
+    NewCqcLocationApiColumns as CQCL,
+)
 from utils.column_values.categorical_column_values import (
+    CQCCurrentOrHistoricValues,
+    CQCRatingsValues,
     IsParent,
+    LocationType,
     ParentsOrSinglesAndSubs,
     RegistrationStatus,
     SingleSubDescription,
@@ -291,4 +297,862 @@ class ReconciliationUtilsData:
     expected_data_for_join_rows = [
         ("loc_1", "estab_1", "name"),
         ("loc_3", "estab_2", None),
+    ]
+
+
+@dataclass
+class FlattenCQCRatings:
+    test_cqc_locations_rows = [
+        (
+            "loc_1",
+            RegistrationStatus.registered,
+            LocationType.social_care_identifier,
+            "20240101",
+            "2024",
+            "01",
+            "01",
+            {
+                CQCL.overall: {
+                    CQCL.organisation_id: None,
+                    CQCL.rating: "Overall rating Excellent",
+                    CQCL.report_date: "report_date",
+                    CQCL.report_link_id: None,
+                    CQCL.use_of_resources: {
+                        CQCL.organisation_id: None,
+                        CQCL.summary: None,
+                        CQCL.use_of_resources_rating: None,
+                        CQCL.combined_quality_summary: None,
+                        CQCL.combined_quality_rating: None,
+                        CQCL.report_date: None,
+                        CQCL.report_link_id: None,
+                    },
+                    CQCL.key_question_ratings: [
+                        {
+                            CQCL.name: "Safe",
+                            CQCL.rating: "Safe rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Well-led",
+                            CQCL.rating: "Well-led rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Caring",
+                            CQCL.rating: "Caring rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Responsive",
+                            CQCL.rating: "Responsive rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Effective",
+                            CQCL.rating: "Effective rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                    ],
+                },
+                CQCL.service_ratings: [
+                    {
+                        CQCL.name: None,
+                        CQCL.rating: None,
+                        CQCL.report_date: None,
+                        CQCL.organisation_id: None,
+                        CQCL.report_link_id: None,
+                        CQCL.key_question_ratings: [
+                            {
+                                CQCL.name: None,
+                                CQCL.rating: None,
+                            },
+                        ],
+                    },
+                ],
+            },
+            [
+                {
+                    CQCL.report_date: "report_date",
+                    CQCL.report_link_id: None,
+                    CQCL.organisation_id: None,
+                    CQCL.service_ratings: [
+                        {
+                            CQCL.name: None,
+                            CQCL.rating: None,
+                            CQCL.key_question_ratings: [
+                                {
+                                    CQCL.name: None,
+                                    CQCL.rating: None,
+                                },
+                            ],
+                        },
+                    ],
+                    CQCL.overall: {
+                        CQCL.rating: "Overall rating Excellent",
+                        CQCL.use_of_resources: {
+                            CQCL.combined_quality_rating: None,
+                            CQCL.combined_quality_summary: None,
+                            CQCL.use_of_resources_rating: None,
+                            CQCL.use_of_resources_summary: None,
+                        },
+                        CQCL.key_question_ratings: [
+                            {CQCL.name: "Safe", CQCL.rating: "Safe rating Good"},
+                            {
+                                CQCL.name: "Well-led",
+                                CQCL.rating: "Well-led rating Good",
+                            },
+                            {CQCL.name: "Caring", CQCL.rating: "Caring rating Good"},
+                            {
+                                CQCL.name: "Responsive",
+                                CQCL.rating: "Responsive rating Good",
+                            },
+                            {
+                                CQCL.name: "Effective",
+                                CQCL.rating: "Effective rating Good",
+                            },
+                        ],
+                    },
+                },
+            ],
+        ),
+    ]
+    test_ascwds_workplace_rows = [("loc_1", "estab_1", "20240101", "2021", "01", "01")]
+    filter_to_first_import_of_most_recent_month_rows = [
+        ("loc_1", "20240101", "2024", "01", "01"),
+        ("loc_2", "20231201", "2023", "12", "01"),
+    ]
+    filter_to_first_import_of_most_recent_month_when_two_imports_in_most_recent_month_rows = [
+        ("loc_1", "20240101", "2024", "01", "01"),
+        ("loc_2", "20231201", "2023", "12", "01"),
+        ("loc_3", "20240104", "2024", "01", "04"),
+    ]
+    filter_to_first_import_of_most_recent_month_when_earliest_date_is_not_first_of_month_rows = [
+        ("loc_1", "20240102", "2024", "01", "02"),
+        ("loc_2", "20231201", "2023", "12", "01"),
+        ("loc_3", "20240104", "2024", "01", "04"),
+    ]
+    expected_filter_to_first_import_of_most_recent_month_rows = [
+        ("loc_1", "20240101", "2024", "01", "01"),
+    ]
+    expected_filter_to_first_import_of_most_recent_month_when_earliest_date_is_not_first_of_month_rows = [
+        ("loc_1", "20240102", "2024", "01", "02"),
+    ]
+
+    flatten_current_ratings_rows = [
+        (
+            "loc_1",
+            "registered",
+            {
+                CQCL.overall: {
+                    CQCL.organisation_id: None,
+                    CQCL.rating: "Overall rating Excellent",
+                    CQCL.report_date: "report_date",
+                    CQCL.report_link_id: None,
+                    CQCL.use_of_resources: {
+                        CQCL.organisation_id: None,
+                        CQCL.summary: None,
+                        CQCL.use_of_resources_rating: None,
+                        CQCL.combined_quality_summary: None,
+                        CQCL.combined_quality_rating: None,
+                        CQCL.report_date: None,
+                        CQCL.report_link_id: None,
+                    },
+                    CQCL.key_question_ratings: [
+                        {
+                            CQCL.name: "Safe",
+                            CQCL.rating: "Safe rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Well-led",
+                            CQCL.rating: "Well-led rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Caring",
+                            CQCL.rating: "Caring rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Responsive",
+                            CQCL.rating: "Responsive rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                        {
+                            CQCL.name: "Effective",
+                            CQCL.rating: "Effective rating Good",
+                            CQCL.report_date: None,
+                            CQCL.organisation_id: None,
+                            CQCL.report_link_id: None,
+                        },
+                    ],
+                },
+                CQCL.service_ratings: [
+                    {
+                        CQCL.name: None,
+                        CQCL.rating: None,
+                        CQCL.report_date: None,
+                        CQCL.organisation_id: None,
+                        CQCL.report_link_id: None,
+                        CQCL.key_question_ratings: [
+                            {
+                                CQCL.name: None,
+                                CQCL.rating: None,
+                            },
+                        ],
+                    },
+                ],
+            },
+        ),
+    ]
+
+    flatten_historic_ratings_rows = [
+        (
+            "loc_1",
+            "registered",
+            [
+                {
+                    CQCL.report_date: "report_date",
+                    CQCL.report_link_id: None,
+                    CQCL.organisation_id: None,
+                    CQCL.service_ratings: [
+                        {
+                            CQCL.name: None,
+                            CQCL.rating: None,
+                            CQCL.key_question_ratings: [
+                                {
+                                    CQCL.name: None,
+                                    CQCL.rating: None,
+                                },
+                            ],
+                        },
+                    ],
+                    CQCL.overall: {
+                        CQCL.rating: "Overall rating Excellent",
+                        CQCL.use_of_resources: {
+                            CQCL.combined_quality_rating: None,
+                            CQCL.combined_quality_summary: None,
+                            CQCL.use_of_resources_rating: None,
+                            CQCL.use_of_resources_summary: None,
+                        },
+                        CQCL.key_question_ratings: [
+                            {CQCL.name: "Safe", CQCL.rating: "Safe rating Good"},
+                            {
+                                CQCL.name: "Well-led",
+                                CQCL.rating: "Well-led rating Good",
+                            },
+                            {CQCL.name: "Caring", CQCL.rating: "Caring rating Good"},
+                            {
+                                CQCL.name: "Responsive",
+                                CQCL.rating: "Responsive rating Good",
+                            },
+                            {
+                                CQCL.name: "Effective",
+                                CQCL.rating: "Effective rating Good",
+                            },
+                        ],
+                    },
+                },
+            ],
+        ),
+    ]
+
+    expected_flatten_ratings_rows = [
+        (
+            "loc_1",
+            "registered",
+            "report_date",
+            "Overall rating Excellent",
+            "Safe rating Good",
+            "Well-led rating Good",
+            "Caring rating Good",
+            "Responsive rating Good",
+            "Effective rating Good",
+        )
+    ]
+    recode_unknown_to_null_rows = [
+        (
+            "loc_1",
+            "registered",
+            "report_date",
+            "Excellent",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_2",
+            "registered",
+            "report_date",
+            "Inspected but not rated",
+            "No published rating",
+            "Insufficient evidence to rate",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_3",
+            "No published rating",
+            "",
+            "Excellent",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_4",
+            "deregistered",
+            "report_date",
+            "Inspected but not rated",
+            "Inspected but not rated",
+            "Inspected but not rated",
+            "Inspected but not rated",
+            "Inspected but not rated",
+            "Inspected but not rated",
+        ),
+        (
+            "loc_4",
+            "deregistered",
+            "report_date",
+            "Inspected but not rated",
+            "Inspected but not rated",
+            "Inspected but not rated",
+            "Inspected but not rated",
+            "Inspected but not rated",
+            "Insufficient evidence to rate",
+        ),
+    ]
+    expected_recode_unknown_to_null_rows = [
+        (
+            "loc_1",
+            "registered",
+            "report_date",
+            "Excellent",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_2",
+            "registered",
+            "report_date",
+            None,
+            None,
+            None,
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_3",
+            "No published rating",
+            "",
+            "Excellent",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_4",
+            "deregistered",
+            "report_date",
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+        ),
+    ]
+
+    add_current_or_historic_rows = [
+        ("loc_1",),
+    ]
+    expected_add_current_rows = [
+        ("loc_1", CQCCurrentOrHistoricValues.current),
+    ]
+    expected_add_historic_rows = [
+        ("loc_1", CQCCurrentOrHistoricValues.historic),
+    ]
+
+    remove_blank_rows_rows = [
+        (
+            "loc_1",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_2",
+            "Registered",
+            "20240101",
+            None,
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_3",
+            "Registered",
+            "20240101",
+            "Good",
+            None,
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_4",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            None,
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_5",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            None,
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_6",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            None,
+            "Good",
+        ),
+        (
+            "loc_7",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            None,
+        ),
+        ("loc_8", "Registered", "20240101", None, None, None, None, None, None),
+        (
+            "loc_1",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+    ]
+    expected_remove_blank_rows_rows = [
+        (
+            "loc_1",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_2",
+            "Registered",
+            "20240101",
+            None,
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_3",
+            "Registered",
+            "20240101",
+            "Good",
+            None,
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_4",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            None,
+            "Good",
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_5",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            None,
+            "Good",
+            "Good",
+        ),
+        (
+            "loc_6",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            None,
+            "Good",
+        ),
+        (
+            "loc_7",
+            "Registered",
+            "20240101",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            None,
+        ),
+    ]
+
+    add_rating_sequence_rows = [
+        ("loc_1", "2024-01-01"),
+        ("loc_1", "2024-01-02"),
+        ("loc_2", "2024-01-01"),
+        ("loc_2", "2024-02-01"),
+        ("loc_3", "2023-01-01"),
+        ("loc_3", "2024-01-01"),
+    ]
+    expected_add_rating_sequence_rows = [
+        ("loc_1", "2024-01-02", 2),
+        ("loc_2", "2024-01-01", 1),
+        ("loc_2", "2024-02-01", 2),
+        ("loc_1", "2024-01-01", 1),
+        ("loc_3", "2023-01-01", 1),
+        ("loc_3", "2024-01-01", 2),
+    ]
+    expected_reversed_add_rating_sequence_rows = [
+        ("loc_1", "2024-01-02", 1),
+        ("loc_2", "2024-01-01", 2),
+        ("loc_2", "2024-02-01", 1),
+        ("loc_1", "2024-01-01", 2),
+        ("loc_3", "2023-01-01", 2),
+        ("loc_3", "2024-01-01", 1),
+    ]
+
+    add_latest_rating_flag_rows = [
+        ("loc_1", 1),
+        ("loc_2", 1),
+        ("loc_2", 2),
+    ]
+    expected_add_latest_rating_flag_rows = [
+        ("loc_1", 1, 1),
+        ("loc_2", 1, 1),
+        ("loc_2", 2, 0),
+    ]
+
+    create_standard_rating_dataset_rows = [
+        (
+            "loc_1",
+            "Registered",
+            "2024-01-01",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Current",
+            1,
+            1,
+            1,
+            3,
+            3,
+            3,
+            3,
+            3,
+            15,
+        ),
+        (
+            "loc_1",
+            "Registered",
+            "2024-01-01",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Current",
+            1,
+            1,
+            1,
+            3,
+            3,
+            3,
+            3,
+            3,
+            15,
+        ),
+        (
+            "loc_1",
+            "Degistered",
+            "2024-01-01",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Current",
+            1,
+            1,
+            1,
+            3,
+            3,
+            3,
+            3,
+            3,
+            15,
+        ),
+        (
+            "loc_1",
+            "Registered",
+            "2024-01-01",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Historic",
+            1,
+            1,
+            1,
+            3,
+            3,
+            3,
+            3,
+            3,
+            15,
+        ),
+        (
+            "loc_1",
+            "Registered",
+            "2024-01-01",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Current",
+            1,
+            0,
+            1,
+            3,
+            3,
+            3,
+            3,
+            3,
+            15,
+        ),
+    ]
+    expected_create_standard_rating_dataset_rows = [
+        (
+            "loc_1",
+            "2024-01-01",
+            "Current",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            1,
+            1,
+            3,
+            3,
+            3,
+            3,
+            3,
+            15,
+        ),
+        (
+            "loc_1",
+            "2024-01-01",
+            "Historic",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            "Good",
+            1,
+            1,
+            3,
+            3,
+            3,
+            3,
+            3,
+            15,
+        ),
+    ]
+    select_ratings_for_benchmarks_rows = [
+        ("loc_1", RegistrationStatus.registered, CQCCurrentOrHistoricValues.current),
+        ("loc_2", RegistrationStatus.registered, CQCCurrentOrHistoricValues.historic),
+        ("loc_3", RegistrationStatus.deregistered, CQCCurrentOrHistoricValues.current),
+        ("loc_4", RegistrationStatus.deregistered, CQCCurrentOrHistoricValues.historic),
+    ]
+    expected_select_ratings_for_benchmarks_rows = [
+        ("loc_1", RegistrationStatus.registered, CQCCurrentOrHistoricValues.current),
+    ]
+
+    add_good_or_outstanding_flag_rows = [
+        ("loc_1", CQCRatingsValues.outstanding),
+        ("loc_2", CQCRatingsValues.good),
+        ("loc_3", "other rating"),
+        ("loc_1", None),
+    ]
+    expected_add_good_or_outstanding_flag_rows = [
+        ("loc_1", CQCRatingsValues.outstanding, 1),
+        ("loc_2", CQCRatingsValues.good, 1),
+        ("loc_3", "other rating", 0),
+        ("loc_1", None, 0),
+    ]
+    ratings_join_establishment_ids_rows = [
+        ("loc_1", "ratings data"),
+        ("loc_3", "ratings data"),
+    ]
+
+    ascwds_join_establishment_ids_rows = [
+        ("loc_1", "estab_1", "20240101"),
+        ("loc_2", "estab_2", "20240101"),
+    ]
+    expected_join_establishment_ids_rows = [
+        ("loc_1", "ratings data", "estab_1"),
+        ("loc_3", "ratings data", None),
+    ]
+    create_benchmark_ratings_dataset_rows = [
+        ("loc_1", "estab_1", 1, "Good", "2024-01-01", ""),
+        ("loc_2", "estab_2", 0, "Requires improvement", "2024-01-01", ""),
+        ("loc_3", None, 1, "Good", "2024-01-01", ""),
+        ("loc_4", "estab_2", 0, None, "2024-01-01", ""),
+        ("loc_5", None, 0, None, "2024-01-01", ""),
+    ]
+    expected_create_benchmark_ratings_dataset_rows = [
+        ("loc_1", "estab_1", 1, "Good", "2024-01-01"),
+        ("loc_2", "estab_2", 0, "Requires improvement", "2024-01-01"),
+    ]
+
+    add_numerical_ratings_rows = [
+        (
+            "loc 1",
+            CQCRatingsValues.good,
+            CQCRatingsValues.outstanding,
+            CQCRatingsValues.requires_improvement,
+            CQCRatingsValues.inadequate,
+            CQCRatingsValues.good,
+            None,
+        ),
+    ]
+    expected_add_numerical_ratings_rows = [
+        (
+            "loc 1",
+            CQCRatingsValues.good,
+            CQCRatingsValues.outstanding,
+            CQCRatingsValues.requires_improvement,
+            CQCRatingsValues.inadequate,
+            CQCRatingsValues.good,
+            None,
+            4,
+            2,
+            1,
+            3,
+            0,
+            10,
+        ),
+    ]
+
+    location_id_hash_rows = [
+        ("1-123",),
+    ]
+    expected_location_id_hash_rows = [
+        ("1-123", "b022a7e5cc45cf3dc578"),
+    ]
+    location_id_hash_ten_digit_rows = [
+        ("1-123456789",),
+    ]
+    expected_location_id_hash_ten_digit_rows = [
+        ("1-123456789", "4a5a7fdc6afede351ffd"),
+    ]
+    location_id_hash_eleven_digit_rows = [
+        ("1-1234567890",),
+    ]
+    expected_location_id_hash_eleven_digit_rows = [
+        ("1-1234567890", "133d74f156c4fba255e9"),
+    ]
+    location_id_hash_twelve_digit_rows = [
+        ("1-12345678901",),
+    ]
+    expected_location_id_hash_twelve_digit_rows = [
+        ("1-12345678901", "cf16d3a6b6648d845fda"),
     ]
